Add files via upload
This project was developed as part of the SHL AI Internship Hiring Assessment 2025, where the goal was to predict grammar proficiency scores based on spoken English audio samples.

The dataset provided by SHL is private and confidential. It contains .wav audio files and associated grammar scores. Therefore, no actual dataset files (audio or CSVs) are included in this repository. Feature extraction and modeling were performed on this data locally.